src/main/python/requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
shapely==2.0.1
onnxruntime==1.20.1
Levenshtein==0.26.1
pyclipper==1.3.0.post6